rag_pipeline/requirements.txt
fastapi
uvicorn

pandas
numpy
faiss-cpu
sentence-transformers
transformers
torch
accelerate